SY-4786: Fold the strided read into iterator.go
The stride is an implementation detail of Iterator.read with no surface of its
own, so it sits with its only caller rather than in a file of its own.
